Prevent tab pane from activating twice
Due to the bindTabToCollapse function, activating a tab activates a panel, and activating a panel activates a tab. This causes double activation which can cause an animation glitch.

Open the demo and click "Key Features" and then "Source" in quick succession. After source loads, key features loads a second time and then source loads a second time. Alternatively, add the "fade" class to the tab panes and watch the content fade in then out then in again.